fix reusability of challenges
By adding the request identifier to the hmac we prevent other the use
of the same challenge for every host and ip. Additionally we now write
the expiration time as pow-random which reduces the amount of data we need
to send to the client. Since everything is inside the hmac anyway we dont
have to be scared of reused challenges anymore

Closes #10