Implement log levels
This groups the emitted messages into customizable log levels allowing
to toggle the amount of output. Structured logging is chosen to allow
for easier parsing of the output.
The change is further a preparation for more detailed debug messages.

The -debug flag is removed.
A flag -loglevel is introduced which accepts any log level supported by
the standard logging library.
A flag -pprof is introduced which enables the pprof listener independent
of the log level.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>